README.md
Quadcopter Dynamic Modeling and Simulation (Quad-Sim) v1.00

Provided here is an assortment of materials designed to assist users in modeling and simulation a quadcopter.
Specifically:
•Test rig designs for component performance measurement
•Several MATLAB data analysis tools and GUIs (R2013a Tested)
•A configurable Simulink quadcopter simulation
•And a bit more stuff…

These materials are partially the result of a Senior Design project at Drexel University. The team consisted of:
D. Hartman, K. Landis, M. Mehrer, S. Moreno, and J. Kim.

As this is our first attempt at a public release of our materials, there are undoubtedly errors, omissions, 
and downright lies contained herein. Expect frequent updates as we find and correct issues.

We do not claim to be experts. All of our materials are provided simply as a service to the 
multi-rotor community in sincere hope that it will prove useful as a basis for further inquiry. Users are 
expected to reference our materials against other more reliable sources, and use their best judgment or 
consult professional advice where appropriate, particularly where safety may be a concern.

Quadcopters and RC vehicles are dangerous and are not toys. 
Use caution and follow all manufacturer safety instructions.

That said, we hope you find these materials helpful. Good luck.

++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
GENERAL INSTRUCTIONS
Contained herein is documentation and instructions pertaining to specific tasks. 
A good starting point is to take a look at what materials are provided within these documents.
In general, it is advisable to add all of the MATLAB and Simulink related folders
to the MATLAB path so that they can be easily accessed within the MATLAB environment.
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
Copyright (C) 2014 D. Hartman, K. Landis, M. Mehrer, S. Moreno, J. Kim

This file is part of a Quadcopter Dynamic Modeling and Simulation package (Quad-Sim).
 
Quad-Sim is free: you can redistribute it and/or modify
it under the terms of the GNU Lesser General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.
 
Quad-Sim is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
 
You should have received a copy of the GNU Lesser General Public License
along with Quad-Sim.  If not, see <http://www.gnu.org/licenses/>.
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++

Thanks,
David